refactor(explorer): rename test project
[skip changelog]
