+++ Libplanet.Explorer.Tests/GraphTypes/AddressTypeTest.cs
@@ -2,7 +2,7 @@ using System;
 using Libplanet.Explorer.GraphTypes;
 using Xunit;
 
-namespace Libplanet.Explorer.UnitTests.GraphTypes
+namespace Libplanet.Explorer.Tests.GraphTypes
 {
     public class AddressTypeTest : ScalarGraphTypeTestBase<AddressType>
     {
